+++ src/Command/Command.php
@@ -5,6 +5,7 @@ namespace Popstas\Transmission\Console\Command;
 use GuzzleHttp;
 use Martial\Transmission\API;
 use Popstas\Transmission\Console;
+use Popstas\Transmission\Console\Application;
 use Popstas\Transmission\Console\Config;
 use Popstas\Transmission\Console\TransmissionClient;
 use Psr\Log\LogLevel;
@@ -16,16 +17,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class Command extends BaseCommand
 {
-    /**
-     * @var Config $config
-     */
-    protected $config;
-
-    /**
-     * @var TransmissionClient $client
-     */
-    private $client;
-
     protected function configure()
     {
         $this->addOption('dry-run', null, InputOption::VALUE_NONE, 'Dry run, don\'t change any data');
@@ -33,53 +24,61 @@ class Command extends BaseCommand
         $this->addOption('config', null, InputOption::VALUE_OPTIONAL, 'Configuration file');
     }
 
+    /**
+     * @return Application
+     */
+    public function getApplication()
+    {
+        return parent::getApplication();
+    }
+
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
-        $logger = $this->getLogger($output);
+        // logger
+        $logger = $this->getApplication()->getLogger();
+        if (!isset($logger)) {
+            $verbosityLevelMap = [
+                LogLevel::NOTICE => OutputInterface::VERBOSITY_NORMAL,
+                LogLevel::INFO   => OutputInterface::VERBOSITY_VERBOSE,
+                LogLevel::DEBUG  => OutputInterface::VERBOSITY_DEBUG,
+            ];
+            $logger = new ConsoleLogger($output, $verbosityLevelMap);
+            $this->getApplication()->setLogger($logger);
+        }
+
+        // config
+        $config = $this->getApplication()->getConfig();
+        if (!isset($config)) {
+            $config = new Config($input->getOption('config'));
+            if ($input->hasOption('host') && $input->getOption('host')) {
+                $config->set('transmission-host', $input->getOption('host'));
+            }
+            $this->getApplication()->setConfig($config);
+        }
+
+        // client
+        $client = $this->getApplication()->getClient();
+        if (!isset($client)) {
+            $this->getApplication()->setClient($this->createClient($config, $logger));
+        }
 
         $logger->info('[{date}] command: {args}', [
             'date' => date('Y-m-d H:i:s'),
             'args' => implode(' ', array_slice($_SERVER['argv'], 1)),
         ]);
 
-        $this->config = new Config();
-        $this->config->loadConfigFile($input->getOption('config'));
-
-        if ($input->hasOption('host') && $input->getOption('host')) {
-            $this->config->set('transmission-host', $input->getOption('host'));
-        }
-
         parent::initialize($input, $output);
     }
 
-    /**
-     * @param OutputInterface $output
-     * @return ConsoleLogger
-     */
-    protected function getLogger(OutputInterface $output)
+    private function createClient(Config $config, ConsoleLogger $logger)
     {
-        $verbosityLevelMap = [
-            LogLevel::NOTICE => OutputInterface::VERBOSITY_NORMAL,
-            LogLevel::INFO   => OutputInterface::VERBOSITY_VERBOSE,
-            LogLevel::DEBUG  => OutputInterface::VERBOSITY_DEBUG,
-        ];
-        return new ConsoleLogger($output, $verbosityLevelMap);
-    }
-
-    protected function getClient(OutputInterface $output)
-    {
-        if (isset($this->client)) {
-            return $this->client;
-        }
-
         $connect = [
-            'host'     => $this->config->get('transmission-host'),
-            'port'     => $this->config->get('transmission-port'),
-            'user'     => $this->config->get('transmission-user'),
-            'password' => $this->config->get('transmission-password'),
+            'host'     => $config->get('transmission-host'),
+            'port'     => $config->get('transmission-port'),
+            'user'     => $config->get('transmission-user'),
+            'password' => $config->get('transmission-password'),
         ];
 
-        $logger = $this->getLogger($output);
         $logger->debug('Connect Transmission using: {user}:{password}@{host}:{port}', $connect);
 
         $base_uri = 'http://' . $connect['host'] . ':' . $connect['port'] . '/transmission/rpc';
@@ -88,8 +87,6 @@ class Command extends BaseCommand
         $api = new API\RpcClient($httpClient, $connect['user'], $connect['password']);
         $api->setLogger($logger);
 
-        $this->client = new TransmissionClient($api);
-
-        return $this->client;
+        return new TransmissionClient($api);
     }
 }